feat: inpakken als installeerbaar Python package (#55)
- Voeg [build-system] toe met hatchling aan pyproject.toml
- Configureer package discovery voor src/ layout
- Verplaats pipeline logica naar src/main.py, root main.py wordt shim
- Voeg CLI entry point toe: studentprognose = "src.main:cli"
- Verplaats pre-commit naar [project.optional-dependencies] dev
- Sluit data/ uit van sdist
