+++ pyproject.toml
@@ -1,3 +1,7 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
 [project]
 name = "studentprognose"
 version = "0.1.0"
@@ -9,9 +13,22 @@ dependencies = [
     "pillow>=12.2.0",
     "openpyxl>=3.1.5",
     "pandas>=2.2.3",
-    "pre-commit>=4.1.0",
     "scikit-learn>=1.6.1",
     "statsmodels>=0.14.4",
     "tqdm>=4.67.1",
     "xgboost>=2.1.3",
 ]
+
+[project.optional-dependencies]
+dev = [
+    "pre-commit>=4.1.0",
+]
+
+[project.scripts]
+studentprognose = "src.main:cli"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src"]
+
+[tool.hatch.build.targets.sdist]
+exclude = ["data/"]
